- removed all uses of BYTE as function return value or functio parameter from the dynamic light code.
These were truncating the actual light values.

+++ src/gl/dynlights/a_dynlight.cpp
@@ -266,7 +266,7 @@ void ADynamicLight::Tick()
 
 	case FlickerLight:
 	{
-		BYTE rnd = randLight();
+		int rnd = randLight();
 		float pct = specialf1 / 360.f;
 		
 		m_currentRadius = float(args[LIGHT_INTENSITY + (rnd >= pct * 255)]);
@@ -293,7 +293,7 @@ void ADynamicLight::Tick()
 	// These need some more work elsewhere
 	case ColorFlickerLight:
 	{
-		BYTE rnd = randLight();
+		int rnd = randLight();
 		float pct = specialf1/360.f;
 		
 		m_currentRadius = m_Radius[rnd >= pct * 255];
